requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
docutils==0.22.3
Flask==3.1.2
flit_core==3.12.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
tomli_w==1.2.0
urllib3==2.5.0
Werkzeug==3.1.3
